requirements.txt
asgiref==3.5.0
boto3==1.21.23
certifi==2021.5.30
chardet==4.0.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==4.0.3
django-environ==0.4.5
django-heroku==0.3.1
django-import-export==2.5.0
django-ipware==3.0.2
django-storages==1.11.1
djangorestframework==3.12.4
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==2.10
jmespath==0.10.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
psycopg2==2.8.6
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-http-client==3.3.7
pytz==2022.1
PyYAML==6.0
requests==2.27.1
sendgrid==6.9.7
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
tablib==3.2.0
urllib3==1.26.9
whitenoise==6.0.0
xlrd==2.0.1
xlwt==1.3.0
